[CODE REVIEW: Refactor]: response dto 패키지 생성
